+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ unreleased
 * Reincorporate the test JSON config loader, active only during tests.
 * Remove the ``DEFAULT`` sentinal value; we'll use 'main' as the default loadable name, just like grandpa used to do. This is a breaking change.
 * Add logging config to the Montague Standard Format.
+* Allow config loaders to skip implementing ``app_config()`` and the like, instead of raising ``NotImplementedError``
 
 0.1.5 (2015-05-12)
 -----------------------------------------